@@ -3,15 +3,15 @@ import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
 import 'package:sqflite/sqflite.dart';
-
 // ignore: implementation_imports
 import 'package:sqflite/src/factory_mixin.dart' as impl;
 import 'package:sqflite/utils/utils.dart';
-import 'package:sqflite_example/model/item.dart';
 import 'package:sqflite_example/src/item_widget.dart';
 import 'package:sqflite_example/utils.dart';
 
 // ignore_for_file: avoid_print
+
+import 'model/item.dart';
 
 /// Manual test page.
 class ManualTestPage extends StatefulWidget {
@@ -19,6 +19,7 @@ class ManualTestPage extends StatefulWidget {
   const ManualTestPage({Key? key}) : super(key: key);
 
   @override
+  // ignore: library_private_types_in_public_api
   _ManualTestPageState createState() => _ManualTestPageState();
 }
 
@@ -45,7 +46,7 @@ class _ManualTestPageState extends State<ManualTestPage> {
     await database!.setVersion(version + 1);
   }
 
-  late List<MenuItem> items;
+  late List<SqfMenuItem> items;
   late List<ItemWidget> itemWidgets;
 
   Future<bool> pop() async {
@@ -55,75 +56,77 @@ class _ManualTestPageState extends State<ManualTestPage> {
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    items = <MenuItem>[
-      MenuItem('SQLite version', () async {
+    items = <SqfMenuItem>[
+      SqfMenuItem('SQLite version', () async {
         final db = await openDatabase(inMemoryDatabasePath);
         final results = await db.rawQuery('select sqlite_version()');
         print('select sqlite_version(): $results');
         var version = results.first.values.first;
         print('sqlite version: $version');
         await db.close();
-        ScaffoldMessenger.of(context).showSnackBar(
-            SnackBar(content: Text('select sqlite_version(): $version')));
+        if (mounted) {
+          ScaffoldMessenger.of(context).showSnackBar(
+              SnackBar(content: Text('select sqlite_version(): $version')));
+        }
       }, summary: 'select sqlite_version()'),
-      MenuItem('Factory information', () async {
+      SqfMenuItem('Factory information', () async {
         var info = databaseFactory.toString();
         print('sqlite database factory: $info');
         ScaffoldMessenger.of(context)
             .showSnackBar(SnackBar(content: Text(info)));
       }, summary: 'toString()'),
-      MenuItem('openDatabase', () async {
+      SqfMenuItem('openDatabase', () async {
         await _openDatabase();
       }, summary: 'Open the database'),
-      MenuItem('BEGIN EXCLUSIVE', () async {
+      SqfMenuItem('BEGIN EXCLUSIVE', () async {
         final db = await _openDatabase();
         await db.execute('BEGIN EXCLUSIVE');
       },
           summary:
               'Execute than exit or hot-restart the application. Open the database if needed'),
-      MenuItem('close', () async {
+      SqfMenuItem('close', () async {
         await _closeDatabase();
       },
           summary:
               'Execute after starting then exit the app using the back button on Android and restart from the launcher.'),
-      MenuItem('delete', () async {
+      SqfMenuItem('delete', () async {
         await _deleteDatabase();
       },
           summary:
               'Try open (then optionally) delete, exit or hot-restart then delete then open'),
-      MenuItem('log level: none', () async {
+      SqfMenuItem('log level: none', () async {
         // ignore: deprecated_member_use
         await Sqflite.devSetOptions(
             // ignore: deprecated_member_use
             SqfliteOptions(logLevel: sqfliteLogLevelNone));
       }, summary: 'No logs'),
-      MenuItem('log level: sql', () async {
+      SqfMenuItem('log level: sql', () async {
         // ignore: deprecated_member_use
         await Sqflite.devSetOptions(
             // ignore: deprecated_member_use
             SqfliteOptions(logLevel: sqfliteLogLevelSql));
       }, summary: 'Log sql command and basic database operation'),
-      MenuItem('log level: verbose', () async {
+      SqfMenuItem('log level: verbose', () async {
         // ignore: deprecated_member_use
         await Sqflite.devSetOptions(
             // ignore: deprecated_member_use
             SqfliteOptions(logLevel: sqfliteLogLevelVerbose));
       }, summary: 'Verbose logs, for debugging'),
-      MenuItem('Get info', () async {
+      SqfMenuItem('Get info', () async {
         final factory = databaseFactory as impl.SqfliteDatabaseFactoryMixin;
         final info = await factory.getDebugInfo();
         print(info.toString());
       }, summary: 'Implementation info (dev only)'),
-      MenuItem('Increment version', () async {
+      SqfMenuItem('Increment version', () async {
         print(await _incrementVersion());
       }, summary: 'Implementation info (dev only)'),
-      MenuItem('Multiple db', () async {
+      SqfMenuItem('Multiple db', () async {
         await Navigator.of(context).push(MaterialPageRoute(builder: (_) {
           return const MultipleDbTestPage();
         }));
       }, summary: 'Open multiple databases'),
       ...[800000, 1500000, 15000000, 150000000]
-          .map((size) => MenuItem('Big blob $size', () async {
+          .map((size) => SqfMenuItem('Big blob $size', () async {
                 await testBigBlog(size);
               }))
     ];
@@ -148,8 +151,10 @@ class _ManualTestPageState extends State<ManualTestPage> {
       }
 
       var ok = (await getValue(id)).length == blob.length;
-      ScaffoldMessenger.of(context)
-          .showSnackBar(SnackBar(content: Text('$size: $ok')));
+      if (mounted) {
+        ScaffoldMessenger.of(context)
+            .showSnackBar(SnackBar(content: Text('$size: $ok')));
+      }
     } finally {
       await db.close();
     }
@@ -162,7 +167,7 @@ class _ManualTestPageState extends State<ManualTestPage> {
               item,
               (item) async {
                 final stopwatch = Stopwatch()..start();
-                final future = (item as MenuItem).run();
+                final future = (item as SqfMenuItem).run();
                 setState(() {});
                 await future;
                 // always add a small delay
@@ -233,6 +238,7 @@ class SimpleDbTestPage extends StatefulWidget {
   final String dbName;
 
   @override
+  // ignore: library_private_types_in_public_api
   _SimpleDbTestPageState createState() => _SimpleDbTestPageState();
 }
 
@@ -281,10 +287,12 @@ class _SimpleDbTestPageState extends State<SimpleDbTestPage> {
               final result =
                   firstIntValue(await db.query('test', columns: ['COUNT(*)']));
               // Temp for nnbd successfull lint
-              ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-                content: Text('$result records'),
-                duration: const Duration(milliseconds: 700),
-              ));
+              if (mounted) {
+                ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+                  content: Text('$result records'),
+                  duration: const Duration(milliseconds: 700),
+                ));
+              }
             }
 
             final items = <Widget>[
